eth balance rounding and moved function into hook
